Capture and Save Single Image
No ROS Compliance yet, but this, built largely off of the "Basic Sample" code for Xeneth SDK, is capable of connecting to the Xenic Gobi camera, grabbing a frame, and saving it to a specified directory.  There seem to be some periodic issues with connecting, or doing the image grab, but it is unclear whether that is a hardware, or software issue.